Fix bug that certain invalid input caused a match
I noticed a bug that I could input eg. "b b b b b b b"
and get results. This is not wanted, though it's not a big deal.

The code so far has been a hack, and I wanted to fix that problem.

The result is a recursive algorithm, which enables a cleaner call
syntax in a list comprehension and better commenting of the code!

And the bug is fixed, of course ;)